+++ app/api/studies/route.ts
@@ -1,7 +1,7 @@
 import { studies } from '@/lib/dummy'
 import { Study } from '@/types/Study'
 
-import { api } from '../utils'
+import { api } from '../utils/factory'
 
 const GET = api.ListFactory<Study>(studies)
 const POST = api.CreateFactory<Study>()
