+++ app/api/study-records/route.ts
@@ -1,7 +1,7 @@
 import { studyRecords } from '@/lib/dummy'
 import { StudyRecord } from '@/types/Study'
 
-import { api } from '../utils'
+import { api } from '../utils/factory'
 
 const GET = api.ListFactory<StudyRecord>(studyRecords)
 const POST = api.CreateFactory<StudyRecord>()
